refactor: update imports to use new entities path
