+++ src/features/pokemon-detail/useStatsGraph.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import * as d3 from "d3";
-import { Stat } from "../../shared/entities";
+import { Stat } from "../../pages/Detail/entities";
 import { graphConfig } from "../../lib/constants";
 
 export const useStatsGraph = (
